Add draft merge/pull request linkifier
This adds code for opening the GitHub or GitLab pull/merge request page
that 'merged' a particular commit.
It uses the respective GitHub and GitLab apis for identifying relevant
pull/merge requests.

I've structured it like the existing fugitive handlers with the idea
that fugitive could call out to the :GBrowse plugins in a similar way to
how it does it for GBrowse.

TODO: Talk to @tpope to see whether there's value in putting this in the fugitive ecosystem.